pyproject.toml
[project]
name = "agentveil"
version = "0.3.0"
description = "Python SDK for Agent Veil Protocol — reputation and identity for AI agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Agent Veil Protocol"}]
keywords = ["ai", "agents", "reputation", "did", "identity", "a2a"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "httpx>=0.27.0",
    "pynacl>=1.5.0",
    "base58>=2.1.0",
]

[project.urls]
Homepage = "https://agentveil.dev"
Documentation = "https://agentveil.dev/docs"
Repository = "https://github.com/creatorrmode-lead/avp-sdk"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[tool.setuptools.packages.find]
include = ["agentveil*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
